refactor: Helpers
- Reclaim winnings button
- Refundable and round canceled hooks
